pyproject.toml
[project]
name = "fabric-learn"
version = "0.1.0"
description = ""
authors = [{ name = "Andy Miles", email = "akmiles@icloud.com" }]
requires-python = ">=3.13"
readme = "README.md"
dependencies = [
    "delta-spark (>=3.3.0,<4.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "jupytext (>=1.16.7,<2.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
]

[dependency-groups]
dev = ["ipykernel>=6.29.5,<7"]

[tool.hatch.build.targets.sdist]
include = ["src/fabric_learn"]

[tool.hatch.build.targets.wheel]
include = ["src/fabric_learn"]

[tool.hatch.build.targets.wheel.sources]
"src/fabric_learn" = "fabric_learn"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
